Extract dependency tracking from Digestor
