+++ actionpack/lib/action_view/digestor.rb
@@ -1,25 +1,8 @@
 require 'thread_safe'
+require 'action_view/dependency_tracker'
 
 module ActionView
   class Digestor
-    EXPLICIT_DEPENDENCY = /# Template Dependency: (\S+)/
-
-    # Matches:
-    #   render partial: "comments/comment", collection: commentable.comments
-    #   render "comments/comments"
-    #   render 'comments/comments'
-    #   render('comments/comments')
-    #
-    #   render(@topic)         => render("topics/topic")
-    #   render(topics)         => render("topics/topic")
-    #   render(message.topics) => render("topics/topic")
-    RENDER_DEPENDENCY = /
-      render\s*                     # render, followed by optional whitespace
-      \(?                           # start an optional parenthesis for the render call
-      (partial:|:partial\s+=>)?\s*  # naming the partial, used with collection -- 1st capture
-      ([@a-z"'][@a-z_\/\."']+)      # the template name itself -- 2nd capture
-    /x
-
     cattr_reader(:cache)
     @@cache = ThreadSafe::Cache.new
 
@@ -47,7 +30,7 @@ module ActionView
     end
 
     def dependencies
-      render_dependencies + explicit_dependencies
+      DependencyTracker.find_dependencies(name, template)
     rescue ActionView::MissingTemplate
       [] # File doesn't exist, so no dependencies
     end
@@ -69,16 +52,16 @@ module ActionView
         name.gsub(%r|/_|, "/")
       end
 
-      def directory
-        name.split("/")[0..-2].join("/")
-      end
-
       def partial?
         false
       end
 
+      def template
+        @template ||= finder.find(logical_name, [], partial?, formats: [ format ])
+      end
+
       def source
-        @source ||= finder.find(logical_name, [], partial?, formats: [ format ]).source
+        template.source
       end
 
       def dependency_digest
@@ -87,26 +70,6 @@ module ActionView
         end
 
         (template_digests + injected_dependencies).join("-")
-      end
-
-      def render_dependencies
-        source.scan(RENDER_DEPENDENCY).
-          collect(&:second).uniq.
-
-          # render(@topic)         => render("topics/topic")
-          # render(topics)         => render("topics/topic")
-          # render(message.topics) => render("topics/topic")
-          collect { |name| name.sub(/\A@?([a-z]+\.)*([a-z_]+)\z/) { "#{$2.pluralize}/#{$2.singularize}" } }.
-
-          # render("headline") => render("message/headline")
-          collect { |name| name.include?("/") ? name : "#{directory}/#{name}" }.
-
-          # replace quotes from string renders
-          collect { |name| name.gsub(/["']/, "") }
-      end
-
-      def explicit_dependencies
-        source.scan(EXPLICIT_DEPENDENCY).flatten.uniq
       end
 
       def injected_dependencies
